modified permissions.py file in the utils directory
