+++ authentication/models.py
@@ -13,9 +13,15 @@ class User(AbstractBaseUser, PermissionsMixin):
     CUSTOMER = 1
     CLIENT = 2
     ADMIN = 3
+    GUEST = 4
+    QUIZ = 5
+    QUIZRESPONSE = 6
     ROLE_CHOICES = (
         (CUSTOMER, 'customer'),
         (CLIENT, 'client'),
         (ADMIN, 'admin'),
+        (GUEST, 'guest'),
+        (QUIZ, 'quiz'),
+        (QUIZRESPONSE, 'quizresponse'),
     )
